Implement TryFrom<proto::SelectionSet> on SelectionSet
More prep work for changing the selection set representation.
